refactor: extract dream memory and input file helpers
Motivation

Reduce two oversized files with low-risk internal extractions so future refactors can proceed incrementally without changing behavior.

Technical Changes

- extract Dream v2 memory parsing, source discovery, curated synthesis, and processed-source persistence into app/services/dream_memory.py
- keep app/services.dream public entry points stable by delegating to the new module and aliasing compatibility exports
- preserve session hashing in app/services/dream.py and restore the local hash helper after extraction
- extract InputBar file accept-string and file-type filtering logic into web/src/components/InputBar.files.ts
- update InputBar to consume the shared helper functions for uploads, paste, drop, and hidden input accept handling
- extend InputBar tests to cover the exported helper functions directly while preserving existing component behavior coverage

Impact

Behavior stays the same while shrinking hotspots and establishing cleaner seams for the next refactor slices.
